Don't scroll to newly inserted rows in TTTableViewController.
In -model:didInsertObject:atIndexPath: of TTTableViewController, the
-scrollToRowAtIndexPath:atScrollPosition:animated: message was send
to the tableView whenever a new row was inserted. This is not going
to work if the new row was inserted as part of a larger change, i.e.
if -beginUpdates was called on the tableView and some new sections
were inserted prior to inserting the row.

In addition, scrolling to arbitrary rows without explicit control
from the App developer isn't good practice either, so just ditch
the scroll code all together.